+++ lib/tapioca.dart
@@ -1,5 +1,0 @@
-export 'package:tapioca_v2/src/cup.dart';
-export 'package:tapioca_v2/src/content.dart';
-export 'package:tapioca_v2/src/tapioca_ball.dart';
-export 'package:tapioca_v2/src/video_editor_process.dart';
-export 'package:tapioca_v2/src/video_editor.dart';