+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## 0.0.1+2
 
-* TODO: Describe initial release.
+* Fix alpha color.
+* Add callback cancel export
+
+## 0.0.1+1
+
+* Initial release(update source code tapioca).